Fix view creation script for Postgres Make view creation consistent for SQLServer Switch to a QueryViewProvider approach to avoid thread-unsafe modification of QueryForms showing provenance data
