+++ api/src/org/labkey/api/query/QueryViewProvider.java
@@ -1,0 +1,13 @@
+package org.labkey.api.query;
+
+import org.jetbrains.annotations.Nullable;
+import org.labkey.api.view.ViewContext;
+import org.springframework.validation.BindException;
+
+public interface QueryViewProvider<FormType>
+{
+    String getDataRegionName();
+
+    @Nullable
+    QueryView createView(ViewContext viewContext, FormType form, BindException errors);
+}
